run() and step() functions are now virtual

+++ PathSolver/src/node.cpp
@@ -4,7 +4,7 @@ Node::Node(float x, float y, int w_size, int dim)
 	:rect_{ sf::RectangleShape() }, type_{ Node::Type::empty_ }, seen_{ false }
 {
 	rect_.setPosition(x, y);
-	rect_.setSize(sf::Vector2f(w_size / dim, w_size / dim)); //Grid is 20x20?
+	rect_.setSize(sf::Vector2f((float)w_size / dim, (float)w_size / dim)); //Grid is 20x20?
 	rect_.setFillColor(sf::Color::White); //set Nodes to white
 	rect_.setOutlineColor(sf::Color::Black);
 	rect_.setOutlineThickness(1);
